Kiểm tra xem người dùng có ở cùng voice channel với bot không

+++ functions/button/B_Lyrics_cancel.js
@@ -21,6 +21,16 @@ module.exports.execute = async ({ interaction, lang }) => {
 		await interaction.message.delete().catch(() => {});
 		return;
 	}
+	// Kiểm tra xem có khóa player không
+	if (queue.metadata.LockStatus && queue.metadata.requestedBy?.id !== interaction.user?.id)
+		return interaction.followUp({ content: lang.until.noPermission, ephemeral: true });
+
+	// Kiểm tra xem người dùng có ở cùng voice channel với bot không
+	const botVoiceChannel = interaction.guild.members.me.voice.channel;
+	const userVoiceChannel = interaction.member.voice.channel;
+	if (!botVoiceChannel || botVoiceChannel.id !== userVoiceChannel?.id)
+		return interaction.followUp({ content: lang.music.NOvoiceMe, ephemeral: true });
+
 	const ZiLyrics = queue.metadata.ZiLyrics;
 	ZiLyrics?.unsubscribe();
 	ZiLyrics.mess.delete().catch(() => {});
